code/tfm_flask/requirements.txt
Flask>=3.0
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
Flask-WTF
WTForms
Flask-Bcrypt
openai-whisper
ollama
cryptography
